refactor: delete spaces in file names
