+++ src/assets/img/australian-dollar-icon.png
