+++ src/assets/img/canadian-dollar-icon.png
