+++ src/assets/img/euro-icon.png
