+++ src/assets/img/yen-icon.png
